Make more AO projects compile against server 11

+++ src/ProSuite.QA.Tests.External/QaExternalServiceBase.cs
@@ -263,7 +263,7 @@ namespace ProSuite.QA.Tests.External
 
 			foreach (InvolvedObjectsMsg involvedTableMsg in involvedObjectsMsg)
 			{
-				foreach (int objectId in involvedTableMsg.ObjectIds)
+				foreach (long objectId in involvedTableMsg.ObjectIds)
 				{
 					result.Add(new InvolvedRow(involvedTableMsg.Dataset.Name, objectId));
 				}
